backend/products/requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6; sys_platform == "win32" or platform_system == "Windows"
fastapi==0.110.2
greenlet==3.0.3; platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64"
h11==0.14.0
httptools==0.6.1
idna==3.7
pydantic==2.7.1
pydantic-core==2.18.2
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0
sniffio==1.3.1
sqlalchemy==2.0.29
starlette==0.37.2
typing-extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0; (sys_platform != "cygwin" and sys_platform != "win32") and platform_python_implementation != "PyPy"
watchfiles==0.21.0
websockets==12.0
